requirements.txt
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.28.21
botocore==1.31.21
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.1.9
django-cors-headers==4.2.0
django-environ==0.10.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
jaraco.classes==3.3.0
jmespath==1.0.1
keyring==24.2.0
libs-exchange-pack @ git+https://github.com/OG-XYZ/libs-exchange-pack
load-dotenv==0.1.0
more-itertools==10.1.0
msgpack==1.0.7
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.2
pika==1.3.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyotp==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-okx==0.2.1
pytz==2023.3
redis==4.6.0
requests==2.31.0
s3transfer==0.6.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
websocket-client==1.5.2
websockets==11.0.3
zipp==3.16.2
zope.interface==6.0
